refactor(utils): 🔧 organize nutrition and calculation utilities and update the corresponding imports
