+++ src/app/dashboard/_sections/data-habits.tsx
@@ -2,7 +2,7 @@ import { Circle, Square } from 'lucide-react'
 import InsightsCard from '../_components/analytics/insights-card'
 import ResumeStreak from '../_components/analytics/resume-streak'
 import { Weights } from '~/types'
-import { calculateNutritionalNeeds } from '~/lib/utils'
+import { calculateNutritionalNeeds } from '~/lib/calculations'
 
 interface Props {
 	userMetadata: UserPublicMetadata
