+++ src/app/food/_components/nutrition-card.tsx
@@ -1,6 +1,6 @@
 import { CakeSlice, EggFried, Flame, Ham } from 'lucide-react'
 import { Card, CardContent, CardHeader, CardTitle } from '~/components/ui/card'
-import { getPercentage, round } from '~/lib/utils'
+import { calculatePercentage, round } from '~/lib/calculations'
 import { NutritionMetrics } from '~/types'
 
 export function NutritionCard({
@@ -31,12 +31,12 @@ export function NutritionCard({
 					{nutrient.needed.toLocaleString()}
 				</div>
 				<p className='text-xs text-gray-500 dark:text-gray-400'>
-					{getPercentage(nutrient)}% of daily goal
+					{calculatePercentage(nutrient)}% of daily goal
 				</p>
 				<div className='mt-4 h-2 overflow-hidden rounded-full dark:bg-primary/20'>
 					<div
 						className='h-full bg-primary'
-						style={{ width: `${getPercentage(nutrient)}%` }}
+						style={{ width: `${calculatePercentage(nutrient)}%` }}
 					></div>
 				</div>
 			</CardContent>
