+++ src/server/utils/nutrition.ts
@@ -1,10 +1,13 @@
 import 'server-only'
 
-import { calculateNutritionalNeeds, getAdjustedDay } from '~/lib/utils'
 import { NutritionMetricsPerDay } from '~/types'
 import { db } from '~/server/db'
 import { consumption, food } from '~/server/db/schema'
 import { eq, and, gte } from 'drizzle-orm'
+import {
+	calculateAdjustedDay,
+	calculateNutritionalNeeds
+} from '~/lib/calculations'
 
 export async function getUserNutritionMetrics(
 	userId: string,
@@ -12,7 +15,7 @@ export async function getUserNutritionMetrics(
 ) {
 	const nutritionMeatrics = calculateNutritionalNeeds(userMetadata)
 	const dayOfWeek = new Date()
-	dayOfWeek.setDate(dayOfWeek.getDate() - getAdjustedDay(dayOfWeek) - 1)
+	dayOfWeek.setDate(dayOfWeek.getDate() - calculateAdjustedDay(dayOfWeek) - 1)
 	const result = await db
 		.select({
 			portion: consumption.portion,
@@ -43,7 +46,7 @@ export async function getUserNutritionMetrics(
 				(Number(portion) / Number(servingSize)) * Number(carbs)
 			const fatsConsumed = (Number(portion) / Number(servingSize)) * Number(fat)
 
-			const day = getAdjustedDay(createdAt)
+			const day = calculateAdjustedDay(createdAt)
 			const nutrition =
 				nutritionMeatricsPerDay[day] ?? structuredClone(nutritionMeatrics)
 
@@ -55,47 +58,4 @@ export async function getUserNutritionMetrics(
 	)
 
 	return nutritionMeatricsPerDay
-}
-
-export function calculateBodyFat({
-	sex,
-	born,
-	height,
-	weights
-}: UserPublicMetadata) {
-	const weightInKg = weights[weights.length - 1]?.value ?? 0
-	const heightInMeters = height * 0.3048
-	const bmi = weightInKg / (heightInMeters * heightInMeters)
-
-	const age = new Date().getFullYear() - parseInt(born, 10)
-
-	let bodyFatPercentage = 1.2 * bmi + 0.23 * age - 16.2
-
-	if (sex === 'female') {
-		bodyFatPercentage = 1.2 * bmi + 0.23 * age - 5.4
-	}
-
-	return bodyFatPercentage
-}
-
-export function calculateGoalProgress({
-	weights,
-	goalWeight
-}: UserPublicMetadata) {
-	const initialWeight = weights[0]?.value ?? 0
-	const currentWeight = weights[weights.length - 1]?.value ?? 0
-
-	if (initialWeight > goalWeight) {
-		const progress =
-			((initialWeight - currentWeight) / (initialWeight - goalWeight)) * 100
-		return progress.toFixed(1)
-	}
-
-	if (initialWeight < goalWeight) {
-		const progress =
-			((currentWeight - initialWeight) / (goalWeight - initialWeight)) * 100
-		return progress.toFixed(1)
-	}
-
-	return 100
 }
